+++ src/main/java/com/marginallyclever/ro3/apps/viewport/viewporttools/move/TranslateToolMulti.java
@@ -57,7 +57,7 @@ public class TranslateToolMulti implements ViewportTool {
         tools.add(toolXZ);
         tools.add(toolYZ);
 
-        texture = Registry.textureFactory.load("/com/marginallyclever/ro3/apps/viewport/renderpasses/navball.png");
+        texture = Registry.textureFactory.load("/com/marginallyclever/ro3/apps/viewport/renderpass/navball.png");
         texture.setDoNotExport(true);
     }
