+++ src/main/resources/com/marginallyclever/ro3/apps/viewport/renderpass/mesh.vert
