+++ src/main/resources/com/marginallyclever/ro3/apps/viewport/renderpass/outline_330.frag
